+++ src/code42cli/cmds/detectionlists/commands.py
@@ -11,6 +11,8 @@ def create_bulk_usage_prefix(detection_list_name):
 
 
 class DetectionListCommandFactory:
+    _USAGE_SUFFIX = u"<username> <optional args>"
+
     def __init__(self, detection_list_name):
         self._name = detection_list_name
         self._usage_prefix = create_usage_prefix(detection_list_name)
@@ -25,9 +27,18 @@ class DetectionListCommandFactory:
 
     def create_add_command(self, handler, arg_customizer):
         return Command(
-            u"add",
+            BulkCommandType.ADD,
             u"Add a user to the {} detection list.".format(self._name),
-            u"{} add <username> <optional args>".format(self._usage_prefix),
+            u"{} {} {}".format(self._usage_prefix, BulkCommandType.ADD, self._USAGE_SUFFIX),
+            handler=handler,
+            arg_customizer=arg_customizer,
+        )
+
+    def create_remove_command(self, handler, arg_customizer):
+        return Command(
+            BulkCommandType.REMOVE,
+            u"Remove a user from the {} detection list.".format(self._name),
+            u"{} {} {}".format(self._usage_prefix, BulkCommandType.REMOVE, self._USAGE_SUFFIX),
             handler=handler,
             arg_customizer=arg_customizer,
         )
@@ -43,11 +54,20 @@ class DetectionListCommandFactory:
 
     def create_bulk_add_command(self, handler):
         return Command(
-            u"add",
+            BulkCommandType.ADD,
             u"Bulk add users to the {} detection list using a csv file.".format(self._name),
-            u"{} add <csv-file>".format(self._bulk_usage_prefix),
+            u"{} {} <csv-file>".format(self._bulk_usage_prefix, BulkCommandType.ADD),
             handler=handler,
             arg_customizer=self._load_bulk_add_description,
+        )
+
+    def create_bulk_remove_command(self, handler):
+        return Command(
+            BulkCommandType.REMOVE,
+            u"Bulk remove users from the {} detection list using a file.".format(self._name),
+            u"{} {} <file>".format(self._bulk_usage_prefix, BulkCommandType.REMOVE),
+            handler=handler,
+            arg_customizer=self._load_bulk_remove_description,
         )
 
     @staticmethod
@@ -60,6 +80,14 @@ class DetectionListCommandFactory:
         csv_file = argument_collection.arg_configs[u"csv_file"]
         csv_file.set_help(
             u"The path to the csv file for bulk adding users to the {} detection list.".format(
+                self._name
+            )
+        )
+
+    def _load_bulk_remove_description(self, argument_collection):
+        users_file = argument_collection.arg_configs[u"users_file"]
+        users_file.set_help(
+            u"A file containing a line-separated list of users to remove form the {} detection list".format(
                 self._name
             )
         )