+++ src/code42cli/cmds/detectionlists/enums.py
@@ -5,9 +5,10 @@ class DetectionLists(object):
 
 class BulkCommandType(object):
     ADD = u"add"
+    REMOVE = u"remove"
 
     def __iter__(self):
-        return iter([self.ADD])
+        return iter([self.ADD, self.REMOVE])
 
 
 class DetectionListUserKeys(object):