+++ src/code42cli/invoker.py
@@ -2,7 +2,11 @@ from __future__ import print_function
 
 import sys
 
+from py42.exceptions import Py42ForbiddenError
+
 from code42cli.parser import ArgumentParserError, CommandParser
+from code42cli.errors import log_error
+from code42cli.util import print_error
 
 
 class CommandInvoker(object):
@@ -19,9 +23,18 @@ class CommandInvoker(object):
             input_args (iter[str]): the full list of arguments
             supplied by the user to `code42` cli command.
         """
-        path_parts = self._get_path_parts(input_args)
-        command = self._commands.get(u" ".join(path_parts))
-        self._try_run_command(command, path_parts, input_args)
+        try:
+            path_parts = self._get_path_parts(input_args)
+            command = self._commands.get(u" ".join(path_parts))
+            self._try_run_command(command, path_parts, input_args)
+        except Py42ForbiddenError as err:
+            log_error(err)
+            print_error(
+                u"You do not have the necessary permissions to perform this task. "
+                u"Try using or creating a different profile."
+            )
+        except Exception as ex:
+            log_error(ex)
 
     def _get_path_parts(self, input_args):
         """Gets the portion of `input_args` that refers to a