+++ tests/cmds/securitydata/test_extraction.py
@@ -12,6 +12,7 @@ from .conftest import (
     get_filter_value_from_json,
     get_test_date_str,
 )
+import code42cli.errors as errors
 
 
 @pytest.fixture
@@ -515,34 +516,41 @@ def test_extract_when_creating_sdk_throws_causes_exit(
         extraction_module.extract(sdk, profile, logger, namespace)
 
 
-def test_extract_when_global_variable_is_true_and_is_interactive_prints_error(
-    sdk, profile, logger, namespace_with_begin, extractor, error_printer, interactive_mode
+def test_extract_when_errored_and_is_interactive_prints_error(
+    mocker, sdk, profile, logger, namespace_with_begin, extractor
 ):
-    extraction_module._EXCEPTIONS_OCCURRED = True
+    errors.ERRORED = False
+    errors_error_printer = mocker.patch("{}.errors.print_error".format(PRODUCT_NAME))
+    errors_interactive_mode = mocker.patch("{}.errors.is_interactive".format(PRODUCT_NAME))
+    errors_interactive_mode.return_value = True
+    errors.ERRORED = True
     extraction_module.extract(sdk, profile, logger, namespace_with_begin)
-    assert error_printer.call_count
+    assert errors_error_printer.call_count
+    errors.ERRORED = False
 
 
-def test_extract_when_global_variable_is_true_and_not_is_interactive_does_not_print_error(
+def test_extract_when_errored_and_is_not_interactive_does_not_print_error(
     sdk, profile, logger, namespace_with_begin, extractor, error_printer, non_interactive_mode
 ):
-    extraction_module._EXCEPTIONS_OCCURRED = True
+    errors.ERRORED = True
     extraction_module.extract(sdk, profile, logger, namespace_with_begin)
     assert not error_printer.call_count
+    errors.ERRORED = False
 
 
-def test_extract_when_global_variable_is_false_and_is_interactive_does_not_print_error(
+def test_extract_when_not_errored_and_is_interactive_does_not_print_error(
     sdk, profile, logger, namespace_with_begin, extractor, error_printer, interactive_mode
 ):
-    extraction_module._EXCEPTIONS_OCCURRED = False
+    errors.ERRORED = False
     extraction_module.extract(sdk, profile, logger, namespace_with_begin)
     assert not error_printer.call_count
+    errors.ERRORED = False
 
 
 def test_when_sdk_raises_exception_global_variable_gets_set(
     mocker, sdk, profile, logger, namespace_with_begin, mock_42
 ):
-    extraction_module._EXCEPTIONS_OCCURRED = False
+    errors.ERRORED = False
     mock_sdk = mocker.MagicMock()
 
     # For ease
@@ -560,4 +568,5 @@ def test_when_sdk_raises_exception_global_variable_gets_set(
     )
 
     extraction_module.extract(sdk, profile, logger, namespace_with_begin)
-    assert extraction_module._EXCEPTIONS_OCCURRED
+    assert errors.ERRORED
+    errors.ERRORED = False